Replace demo/colors.rs with a shell for interactive testing
Previous version saved as colors_old.rs.
