+++ demo/README.md
@@ -443,12 +443,28 @@ which will give identical output to the above.
 
 ### Script: colors.rs
 
-**Description:**  A version of `thag_rs`'s `colors` module to style messages according to their type. Like the `stdin`
- module, `colors` was originally developed here as a separate script and integrated as a module later.
-
- Format: `nu_color_println!(style: Option<Style>, "Lorem ipsum dolor {} amet", content: &str);`
+**Description:**  Runner for current version of `src/colors.rs`, as it's become too enmeshed with other modules to split out nicely.
+ We just borrow the main method here and add all the necessary dependencies and imports.
 
  E.g. `thag demo/colors.rs`
+ Main function for use by testing or the script runner.
+
+**Purpose:** Test the look of the various colours.
+
+**Crates:** `nu_ansi_term`, `strum`, `termbg`, `thag_rs`
+
+**Type:** Program
+
+**Link:** [colors.rs](https://github.com/durbanlegend/thag_rs/blob/master/demo/colors.rs)
+
+---
+
+### Script: colors_old.rs
+
+**Description:**  An older version of `thag_rs`'s `colors` module to style messages according to their type. Like the `stdin`
+ module, `colors` was originally developed here as a separate script and integrated as a module later.
+
+ E.g. `thag demo/colors_old.rs`
 
 **Purpose:** Demo using `thag_rs` to develop a module outside of the project.
 
@@ -456,7 +472,7 @@ which will give identical output to the above.
 
 **Type:** Program
 
-**Link:** [colors.rs](https://github.com/durbanlegend/thag_rs/blob/master/demo/colors.rs)
+**Link:** [colors_old.rs](https://github.com/durbanlegend/thag_rs/blob/master/demo/colors_old.rs)
 
 ---
 
@@ -466,7 +482,6 @@ which will give identical output to the above.
  to their type. I only dropped `owo-colors` because I switched from `rustyline` to
  `reedline`, which was already using `nu_ansi_term`.
 
- Format: `color_println!(style: Option<Style>, "Lorem ipsum dolor {} amet", content: &str);`
 
 **Purpose:** Demo older alternative implementation of `colors` module using `owo-colors`.
 
@@ -770,6 +785,21 @@ which will give identical output to the above.
 **Type:** Program
 
 **Link:** [darling_consume_fields.rs](https://github.com/durbanlegend/thag_rs/blob/master/demo/darling_consume_fields.rs)
+
+---
+
+### Script: darling_struct.rs
+
+**Description:**  Published example from `darling` crate showing parsing for derive input.
+ Extended to show formatted version of emitted code.
+
+**Purpose:** Explore `darling` crate.
+
+**Crates:** `darling`, `proc_macro2`, `quote`, `syn`
+
+**Type:** Program
+
+**Link:** [darling_struct.rs](https://github.com/durbanlegend/thag_rs/blob/master/demo/darling_struct.rs)
 
 ---
 
@@ -2909,7 +2939,7 @@ which will give identical output to the above.
 
 **Purpose:** Allow checking of terminals on platforms to be supported, also test reliability of different crates.
 
-**Crates:** `crossterm`, `supports_color`
+**Crates:** `crossterm`, `log`, `simplelog`, `supports_color`, `thag_rs`
 
 **Type:** Snippet
